+++ src/pretalx/cfp/views/user.py
@@ -15,6 +15,7 @@ from django.views.generic import (
     DetailView, FormView, ListView, TemplateView, UpdateView, View,
 )
 from django_context_decorator import context
+from rest_framework.authtoken.models import Token
 
 from pretalx.cfp.forms.submissions import SubmissionInvitationForm
 from pretalx.cfp.views.event import LoggedInEventPageMixin
@@ -38,6 +39,12 @@ class ProfileView(LoggedInEventPageMixin, TemplateView):
             user=self.request.user,
             data=self.request.POST if is_form_bound(self.request, 'login') else None,
         )
+
+    @context
+    def token(self):
+        return Token.objects.filter(
+            user=self.request.user
+        ).first() or Token.objects.create(user=self.request.user)
 
     @context
     @cached_property
@@ -70,7 +77,16 @@ class ProfileView(LoggedInEventPageMixin, TemplateView):
         return self.request.event.questions.filter(target='speaker').exists()
 
     def post(self, request, *args, **kwargs):
-        if self.login_form.is_bound and self.login_form.is_valid():
+        if request.POST.get('form') == 'token':
+            request.user.regenerate_token()
+            messages.success(
+                request,
+                _(
+                    'Your API token has been regenerated. The previous token will not be usable any longer.'
+                ),
+            )
+            return super().get(request, *args, **kwargs)
+        elif self.login_form.is_bound and self.login_form.is_valid():
             self.login_form.save()
             request.user.log_action('pretalx.user.password.update')
         elif self.profile_form.is_bound and self.profile_form.is_valid():